Split testlib into multiple modules
